lr is not int its float

+++ MachineTranslation/Train.py
@@ -129,5 +129,5 @@ if __name__ == "__main__":
     if (sys.argv.__len__() < 6):
         model = train("dataset/train_p", "dataset/dev_p", 128, 10, 5, lr=1)
     else:
-        model = train(sys.argv[1], sys.argv[2], 128, 10, 5, lr=int(sys.argv[3]),
+        model = train(sys.argv[1], sys.argv[2], 128, 10, 5, lr=float(sys.argv[3]),
                       model_save_dir=sys.argv[4], model_save_filename=sys.argv[5])
